+++ editor/editor-common/src/utils/breakout.ts
@@ -6,71 +6,73 @@ import {
   breakoutWideScaleRatio,
 } from '@atlaskit/editor-shared-styles';
 
-import { mapBreakpointToLayoutMaxWidth as mapBreakpointToLayoutMaxWidthExported } from '../ui/BaseTheme';
-import { getBreakpoint as getBreakpointExported } from '../ui/WidthProvider';
+import { mapBreakpointToLayoutMaxWidth } from '../ui/BaseTheme';
+import { getBreakpoint } from '../ui/WidthProvider';
 
 /**
- * Variables required to get better naming and
- * cleaner context for breakout ssr inline script.
- *
- * E.g.:
- * – getBreakpoint vs ModuleName.getBreakpoint
- * – breakoutConsts.padding vs consts_1.akEditorBreakoutPadding
+ * Variables required to construct a context for breakout ssr inline script.
  *
  * TODO: Clean this up after: https://product-fabric.atlassian.net/browse/ED-8942
  */
-const mapBreakpointToLayoutMaxWidth = mapBreakpointToLayoutMaxWidthExported;
-const getBreakpoint = getBreakpointExported;
-const breakoutConsts = {
+const breakoutConsts: any = {
   padding: akEditorBreakoutPadding,
   defaultLayoutWidth: akEditorDefaultLayoutWidth,
   wideScaleRatio: breakoutWideScaleRatio,
   fullWidthLayoutWidth: akEditorFullWidthLayoutWidth,
   wideLayoutWidth: akEditorWideLayoutWidth,
-};
+  mapBreakpointToLayoutMaxWidth,
+  getBreakpoint,
+  calcBreakoutWidth: (
+    layout: 'full-width' | 'wide' | string,
+    containerWidth: number,
+  ) => {
+    const effectiveFullWidth = containerWidth - breakoutConsts.padding;
 
-const calcBreakoutWidth = (
-  layout: 'full-width' | 'wide' | string,
-  containerWidth: number,
-) => {
-  const effectiveFullWidth = containerWidth - breakoutConsts.padding;
-
-  switch (layout) {
-    case 'full-width':
-      return `${Math.min(
-        effectiveFullWidth,
-        breakoutConsts.fullWidthLayoutWidth,
-      )}px`;
-    case 'wide':
-      return calcWideWidth(containerWidth);
-    default:
-      return '100%';
-  }
-};
-
-const calcWideWidth = (
-  containerWidth: number = breakoutConsts.defaultLayoutWidth,
-  maxWidth: number = Infinity,
-  fallback: string = '100%',
-) => {
-  const effectiveFullWidth = containerWidth - breakoutConsts.padding;
-  const layoutMaxWidth = mapBreakpointToLayoutMaxWidth(
-    getBreakpoint(containerWidth),
-  );
-  const wideWidth = Math.min(
-    Math.ceil(layoutMaxWidth * breakoutConsts.wideScaleRatio),
-    effectiveFullWidth,
-  );
-  return layoutMaxWidth > wideWidth
-    ? fallback
-    : `${Math.min(maxWidth, wideWidth)}px`;
-};
+    switch (layout) {
+      case 'full-width':
+        return `${Math.min(
+          effectiveFullWidth,
+          breakoutConsts.fullWidthLayoutWidth,
+        )}px`;
+      case 'wide':
+        return breakoutConsts.calcWideWidth(containerWidth);
+      default:
+        return '100%';
+    }
+  },
+  calcLineLength: (containerWidth?: number, allowDynamicTextSizing?: boolean) =>
+    allowDynamicTextSizing && containerWidth
+      ? breakoutConsts.mapBreakpointToLayoutMaxWidth(
+          breakoutConsts.getBreakpoint(containerWidth),
+        )
+      : breakoutConsts.defaultLayoutWidth,
+  calcWideWidth: (
+    containerWidth: number = breakoutConsts.defaultLayoutWidth,
+    maxWidth: number = Infinity,
+    fallback: string = '100%',
+  ) => {
+    const effectiveFullWidth = containerWidth - breakoutConsts.padding;
+    const layoutMaxWidth = breakoutConsts.mapBreakpointToLayoutMaxWidth(
+      breakoutConsts.getBreakpoint(containerWidth),
+    );
+    const wideWidth = Math.min(
+      Math.ceil(layoutMaxWidth * breakoutConsts.wideScaleRatio),
+      effectiveFullWidth,
+    );
+    return layoutMaxWidth > wideWidth
+      ? fallback
+      : `${Math.min(maxWidth, wideWidth)}px`;
+  },
+} as const;
 
 export const absoluteBreakoutWidth = (
   layout: 'full-width' | 'wide' | string,
   containerWidth: number,
 ) => {
-  const breakoutWidth = calcBreakoutWidth(layout, containerWidth);
+  const breakoutWidth = breakoutConsts.calcBreakoutWidth(
+    layout,
+    containerWidth,
+  );
 
   // If it's percent, map to max layout size
   if (breakoutWidth.endsWith('%')) {
@@ -80,11 +82,15 @@ export const absoluteBreakoutWidth = (
       case 'wide':
         return akEditorWideLayoutWidth;
       default:
-        return mapBreakpointToLayoutMaxWidth(getBreakpoint(containerWidth));
+        return breakoutConsts.mapBreakpointToLayoutMaxWidth(
+          breakoutConsts.getBreakpoint(containerWidth),
+        );
     }
   }
 
   return parseInt(breakoutWidth, 10);
 };
 
-export { calcBreakoutWidth, calcWideWidth, breakoutConsts };
+export { breakoutConsts };
+export const calcWideWidth = breakoutConsts.calcWideWidth;
+export const calcBreakoutWidth = breakoutConsts.calcBreakoutWidth;